+++ Model/Attribute/Source/AdminConfirmationTemplate.php
@@ -1,0 +1,30 @@
+<?php
+/**
+ * Copyright © Alekseon sp. z o.o.
+ * http://www.alekseon.com/
+ */
+namespace Alekseon\CustomFormsEmailNotification\Model\Attribute\Source;
+
+use Alekseon\AlekseonEav\Model\Attribute\Source\EmailTemplate;
+
+/**
+ *
+ */
+class AdminConfirmationTemplate extends EmailTemplate
+{
+    /**
+     * @return array|mixed|void
+     */
+    public function getOptions()
+    {
+        $options = [
+            0 => __('From Configuration'),
+        ];
+        $templateOptions = parent::getOptions();
+        foreach ($templateOptions as $value => $label) {
+            $options[$value] = $label;
+        }
+
+        return $options;
+    }
+}