+++ Model/Email/CustomerConfirmation.php
@@ -5,145 +5,52 @@
  */
 namespace Alekseon\CustomFormsEmailNotification\Model\Email;
 
-use Alekseon\CustomFormsBuilder\Model\FormRecord;
-use Magento\Framework\App\Config\ScopeConfigInterface;
-use Magento\Framework\Mail\Template\SenderResolverInterface;
-use Magento\Framework\Mail\Template\TransportBuilder;
-use Magento\Store\Model\ScopeInterface;
-use Magento\Store\Model\StoreManagerInterface;
-use Psr\Log\LoggerInterface;
-
 /**
  * Class CustomerConfirmation
  * @package Alekseon\CustomFormsEmailNotification\Model\Email
  */
-class CustomerConfirmation
+class CustomerConfirmation extends AbstractSender
 {
     /**
-     * @var TransportBuilder
+     * @return array|false|string[]
      */
-    protected TransportBuilder $transportBuilder;
-    /**
-     * @var StoreManagerInterface
-     */
-    protected StoreManagerInterface $storeManager;
-    /**
-     * @var LoggerInterface
-     */
-    protected LoggerInterface $logger;
-
-    /**
-     * @var SenderResolverInterface
-     */
-    protected SenderResolverInterface $senderResolver;
-
-    /**
-     * @var ScopeConfigInterface
-     */
-    protected ScopeConfigInterface $scopeConfig;
-
-    /**
-     * EmailNotification constructor.
-     * @param TransportBuilder $transportBuilder
-     * @param StoreManagerInterface $storeManager
-     * @param SenderResolverInterface $senderResolver
-     * @param ScopeConfigInterface $scopeConfig
-     * @param LoggerInterface $logger
-     */
-    public function __construct(
-        TransportBuilder $transportBuilder,
-        StoreManagerInterface $storeManager,
-        SenderResolverInterface $senderResolver,
-        ScopeConfigInterface $scopeConfig,
-        LoggerInterface $logger
-    )
+    public function getReceiverEmails()
     {
-        $this->transportBuilder = $transportBuilder;
-        $this->storeManager = $storeManager;
-        $this->senderResolver = $senderResolver;
-        $this->scopeConfig = $scopeConfig;
-        $this->logger = $logger;
-    }
-
-    /**
-     * @param FormRecord $formRecord
-     */
-    public function send(FormRecord $formRecord)
-    {
-        $form = $formRecord->getForm();
-
-        if (!$form->getCustomerEmailNotificationEnable()) {
-            return false;
-        }
-
+        $receiverEmails = [];
+        $form = $this->formRecord->getForm();
         $emailField = $form->getCustomerNotificationEmailField();
-        if (!$emailField) {
-            return false;
+        if ($emailField) {
+            $email = $this->formRecord->getData((string)$emailField);
+            if ($email) {
+                $receiverEmails[] = $email;
+            }
         }
 
-        $email = $formRecord->getData((string) $emailField);
-
-        if (!$email) {
-            return false;
+        if (!empty($receiverEmails)) {
+            if ($form->getCustomerNotificationCopyTo()) {
+                $copyTo = explode(',', $form->getCustomerNotificationCopyTo());
+                $receiverEmails = array_merge($copyTo, $receiverEmails);
+            }
         }
 
-        $templateId = $form->getCustomerNotificationTemplate();
-        $sender = $form->getCustomerNotificationIdentity();
-
-        $emails = [$email];
-        if ($form->getCustomerNotificationCopyTo()) {
-            $copyTo = explode(',', $form->getCustomerNotificationCopyTo());
-            $emails = array_merge($copyTo, $emails);
-        }
-
-        $templateParams = [
-            'form' => $form,
-            'record' => $formRecord,
-        ];
-
-        return $this->sendEmailTemplate($emails, $templateId, $sender, $templateParams);
+        return $receiverEmails;
     }
 
     /**
-     * @param $email
-     * @param $templateId
-     * @return bool
-     * @throws \Magento\Framework\Exception\LocalizedException
+     * @return mixed
      */
-    protected function sendEmailTemplate($emails, $templateId,$sender, $templateParams = [])
+    public function getSender()
     {
-        $email = array_pop($emails);
+        $form = $this->formRecord->getForm();
+        return $form->getCustomerNotificationIdentity();
+    }
 
-        if (!$email) {
-            return false;
-        }
-
-        $from = $this->senderResolver->resolve($sender);
-        $storeId = $this->storeManager->getDefaultStoreView()->getId();
-
-        $this->transportBuilder->setTemplateIdentifier($templateId)
-            ->setTemplateOptions(
-                [
-                    'area' => \Magento\Framework\App\Area::AREA_ADMINHTML,
-                    'store' => $storeId
-                ]
-            )
-            ->setTemplateVars($templateParams)
-            ->setFrom($from)
-            ->addTo($email);
-
-        foreach ($emails as $email) {
-            $this->transportBuilder->addBcc($email);
-        }
-
-        $transport = $this->transportBuilder->getTransport();
-
-        try {
-            $transport->sendMessage();
-            return true;
-        } catch ( \Exception $e) {
-            $this->logger->critical($e->getMessage());
-            return false;
-        }
+    /**
+     * @return mixed
+     */
+    public function getTemplateId()
+    {
+        $form = $this->formRecord->getForm();
+        return $form->getCustomerNotificationTemplate();
     }
 }